Fix glass int->byte

+++ src/MiNET/MiNET/Blocks/Glass.cs
@@ -6,7 +6,7 @@
 		{
 			
 		}
-		public Glass(int id) : base(id)
+		public Glass(byte id) : base(id)
 		{
 			IsTransparent = true;
 			BlastResistance = 1.5f;
